# PERF: Eliminate dependency on mime-types gem