--- a/README.md
+++ b/README.md
@@ -37,7 +37,7 @@ updates.
 
 The rest-client gem depends on these other gems for usage at runtime:
 
-* [mini-mime](http://rubygems.org/gems/mini-mime)
+* [mini_mime](http://rubygems.org/gems/mini_mime)
 * [netrc](http://rubygems.org/gems/netrc)
 * [http-cookie](https://rubygems.org/gems/http-cookie)