--- a/lib/restclient/request.rb
+++ b/lib/restclient/request.rb
@@ -1,5 +1,5 @@
 require 'tempfile'
-require 'mime/types'
+require 'mini_mime'
 require 'cgi'
 require 'netrc'
 require 'set'
@@ -912,12 +912,8 @@ module RestClient
         return ext
       end
 
-      types = MIME::Types.type_for(ext)
-      if types.empty?
-        ext
-      else
-        types.first.content_type
-      end
+      type = MiniMime.lookup_by_filename("a.#{ext}")
+      type ? type.content_type : ext
     end
   end
 end