--- a/rest-client.gemspec
+++ b/rest-client.gemspec
@@ -24,7 +24,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency('rubocop', '~> 0')
 
   s.add_dependency('http-cookie', '>= 1.0.2', '< 2.0')
-  s.add_dependency('mime-types', '>= 1.16', '< 4.0')
+  s.add_dependency('mini_mime', '>= 0.1.1')
   s.add_dependency('netrc', '~> 0.8')
 
   s.required_ruby_version = '>= 2.0.0'