@ioquatix 2016-12-14
If it's any consolation, I'm working on a gem that does exactly the same thing, with almost the same API but only uses about 300Kbytes of memory. It does it because it compiles it all to C rather than using an in-memory hash table. IF this is of interested, I can accelerate the speed of development on this project, although I do have a proof of concept and it works well.