@SamSaffron 2016-12-14
@ioquatix Nice work! I think I came across this, the challenge with Rails though is that there is a strong preference to depend on gems without c extensions and this dependency is baked in super deep...

action mailer -> mail -> mime types

I think you should open some discussion with the mime types team to see if they are interested in integrating your work so a gem can be installed that picks best implementation based on available VM. 

That said, at a much higher level the majority of what the mime types gem stores is unneeded information for the 2 giant consumers of the gem in the wild (mail and rest-client) I think full featured "mime types" is unneeded for 99% of the consumers, no need to ship all this complex logic and so on for the general case.