@ioquatix 2016-12-14
@SamSaffron I am actually in communication with the Mime-Types developer and we've discussed this for a while.https://github.com/mime-types/ruby-mime-types/issues/123