@ioquatix 2016-12-14
In particular see https://github.com/mime-types/ruby-mime-types/issues/123#issuecomment-252483909