@SamSaffron 2017-01-11
@ab any news here? this was merged into the mail gem recently