@ioquatix 2017-01-11
@SamSaffron What was merged into the mail gem?