@SamSaffron 2017-01-11
@ioquatix  

https://github.com/mikel/mail/commit/634d91145cb65a85dff9da61eddc9e2de0609f8b