@ioquatix 2017-01-11
I haven't had a close look at how this can be used, but I think I have a problem with the overall design.

It should be possible to plug in mime type libraries - it's a super simple feature. Is this possible with mini_mime?