@SamSaffron 2017-01-11
If we want to move to a execjs like design for mime types it is totally
doable, but someone is going to have to make an execjs like gem, maintain
it, test it, design the interfaces and so on. this is a fair amount of work
to take on imo despite the small amount of code.

On Wed, Jan 11, 2017 at 12:59 AM, Samuel Williams <notifications@github.com>
wrote:

> I haven't had a close look at how this can be used, but I think I have a
> problem with the overall design.
>
> It should be possible to plug in mime type libraries - it's a super simple
> feature. Is this possible with mini_mime?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/rest-client/rest-client/pull/557#issuecomment-271785542>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAAUXZGHG7zayS8BNBFFBUxT1ZF000Mgks5rRG_VgaJpZM4LMie6>
> .
>