@Fryguy 2017-03-27
Bumping to see if this can get reviewed / merged?  It's quite a huge improvement, and when combined with the [already merged PR in the mail gem](https://github.com/mikel/mail/pull/1060), we can eliminate the mime-types gem entirely, for a gain of about ~10MB per per process (at least in production...I'm seeing capybara still using mime-types for some reason in test/dev).

I understand the suggestion to have an abstraction layer over mime type libraries, but I wonder if that can be done in follow up PRs so as not to hold up this PR.