@mwpastore 2017-11-17
mail v2.7.0 was released on Nerd Christmas, officially switching from mime-types to mimi_mime. Your move, rest-client maintainers!