@Fryguy 2017-11-20
Created rebased replacement PR #644